[bug #165645049] this commit fixed the bug that not integer account can be debited or credited and rearrange the credit and debit files
