+++ server/controllers/accounts/activate.deactivateAcc.js
@@ -21,8 +21,8 @@ const patch = {
         }
         // verify if the account activate or deactive exist
         if ((isNaN(req.params.accountNumber))) {
-          return res.status(400).json({
-            status: 400,
+          return res.status(403).json({
+            status: 403,
             message: 'Sorry the account number do not exist or is not an integer',
           });
         }
