+++ server/controllers/accounts/view.specific.accounts.js
@@ -25,8 +25,8 @@ class ViewAccounts {
       }
       // Verify is the account is an integer
       if ((isNaN(req.params.accountNumber, 10))) {
-        return res.status(400).json({
-          status: 400,
+        return res.status(403).json({
+          status: 403,
           message: 'Sorry the account number do not exist or is not an integer',
         });
       }
